Move common getUpdatedState logic to StreamBase getUpdatedStreamState function. Fix schema and imports.
