+++ sources/jira-source/src/jira.ts
@@ -4,8 +4,7 @@ import {AirbyteConfig, AirbyteLogger} from 'faros-airbyte-cdk';
 import {Utils, wrapApiError} from 'faros-js-client';
 import parseGitUrl from 'git-url-parse';
 import https from 'https';
-import jira from 'jira.js';
-import {AgileModels, Version2Models} from 'jira.js';
+import jira, {AgileModels, Version2Models} from 'jira.js';
 import {concat, isNil, sum, toLower} from 'lodash';
 import pLimit from 'p-limit';
 import {Memoize} from 'typescript-memoize';
